+++ zallet/src/components/json_rpc/methods/z_send_many.rs
@@ -50,8 +50,8 @@ use crate::{
     prelude::*,
 };
 
-#[cfg(feature = "transparent-key-import")]
-use {transparent::address::TransparentAddress, zcash_script::script};
+#[cfg(feature = "zcashd-import")]
+use crate::components::json_rpc::utils::collect_standalone_transparent_keys;
 
 #[derive(Serialize, Deserialize, JsonSchema)]
 pub(crate) struct AmountParameter {
@@ -363,34 +363,10 @@ pub(crate) async fn call(
     )
     .map_err(|e| LegacyCode::InvalidAddressOrKey.with_message(e.to_string()))?;
 
-    #[cfg(feature = "transparent-key-import")]
-    let standalone_keys = {
-        let mut keys = std::collections::HashMap::new();
-        for step in proposal.steps() {
-            for input in step.transparent_inputs() {
-                if let Some(address) = script::FromChain::parse(&input.txout().script_pubkey().0)
-                    .ok()
-                    .as_ref()
-                    .and_then(TransparentAddress::from_script_from_chain)
-                {
-                    let secret_key = keystore
-                        .decrypt_standalone_transparent_key(&address)
-                        .await
-                        .map_err(|e| match e.kind() {
-                            // TODO: Improve internal error types.
-                            crate::error::ErrorKind::Generic
-                                if e.to_string() == "Wallet is locked" =>
-                            {
-                                LegacyCode::WalletUnlockNeeded.with_message(e.to_string())
-                            }
-                            _ => LegacyCode::Database.with_message(e.to_string()),
-                        })?;
-                    keys.insert(address, vec![secret_key]);
-                }
-            }
-        }
-        keys
-    };
+    #[cfg(feature = "zcashd-import")]
+    let standalone_keys =
+        collect_standalone_transparent_keys(wallet.as_ref(), &keystore, account.id(), &proposal)
+            .await?;
 
     // TODO: verify that the proposal satisfies the requested privacy policy
 
@@ -407,11 +383,10 @@ pub(crate) async fn call(
             wallet,
             chain,
             proposal,
-            SpendingKeys::new(
-                usk,
-                #[cfg(feature = "zcashd-import")]
-                standalone_keys,
-            ),
+            #[cfg(feature = "zcashd-import")]
+            SpendingKeys::new(usk, standalone_keys),
+            #[cfg(not(feature = "zcashd-import"))]
+            SpendingKeys::from_unified_spending_key(usk),
         ),
     ))
 }